bugfix: revoke message was not sent when a (currently) utterance-final word is replaced by a silence word.
the relevant part of the diff is the following:
Index: src/org/cocolab/inpro/incremental/deltifier/ASRWordDeltifier.java
===================================================================
--- src/org/cocolab/inpro/incremental/deltifier/ASRWordDeltifier.java	(revision 1842)
+++ src/org/cocolab/inpro/incremental/deltifier/ASRWordDeltifier.java	(working copy)
@@ -151,11 +194,18 @@
 				if (currSegmentIt.hasNext()) {
 					currSegmentIt.next().updateLabel(new Label(segmentStartTime, segmentEndTime, name));
 				} else if (name.equals("SIL")) {
-					if (currWordIt.hasNext() && currWordIt.next().isSilence()) {
-						currWordIt.previous();
-						WordIU currWord = currWordIt.next(); 
-						currWord.updateSegments(Collections.nCopies(1, new Label(segmentStartTime, segmentEndTime, name)));
-						currSegmentIt = currWord.getSegments().iterator();
+					if (wordIt.hasNext()) {
+						if (wordIt.next().isSilence()) {
+							wordIt.previous();
+							WordIU currWord = wordIt.next(); 
+							currWord.updateSegments(Collections.nCopies(1, new Label(segmentStartTime, segmentEndTime, name)));
+							currSegmentIt = currWord.getSegments().iterator();
+							addSilenceWord = false; // maybe this was missing? -> nope, doesn't help, but still, it seems sane to be here
+						} else {
+							wordIt.previous();
+							addSilenceWord = true;
+							break;
+						}
 					} else {
 						addSilenceWord = true;
 						break;



The problem here is the "complex" condition of the if-statement. 
We actually have to distinguish whether the first part of the condition (wordIt.hasNext()) fails, or whether wordIt.next().isSilence() fails. The problem is that wordIt.next() has to be "undone" because we need this word to be returned by the iterator later on. Now, if the first part of the condition holds, wordIt.next() has been called, but if the first part of the condition does not hold, wordIt.next() has not been called. Hence, the two cases have to be distinguished. Before this patch, wordIt.previous() is only called if both conditions hold. As a consequence, a REVOKE for the (currently) utterance-final word is not sent if it is to be replaced by a silence.

git-svn-id: svn+ssh://gate.spectrum.uni-bielefeld.de/homes/tbaumann/inpro-svnrepo/trunk@1845 19460358-522e-0410-942e-fdb976ae3ca2