fix CarChaseExperimenter's Hesitation implementation
Hesitations in CarChaseExperimenter broke with the introduction of a syllable layer
in speech synthesis (commit. a09b66bd70), but this was not caught as the test for
CarChase did not trigger hesitations. Here, we change the config that is synthesized
by default (thus, triggering the hesitation case in tests) and fix
HesitatingSynthesisInstallment (and its implementation of HesitationIU) to account
for SyllableIUs. Also, rename TTSAction.appData to much more precise installmentIU.