test to ensure that synthesis takes some time
test that some time passes between the calls to SysSegmentIU#update()
which is an indicator that synthesis is actually running in real-time,
instead of an audio stream being drained in no time and later (possibly)
played in real-time. In that case, there would be no correspondence
between updates and what's on the speakers.